requirements.txt
Python 
Transformers
PyTorch
scikit-learn
pandas